Include QPlainTextEdit with no file extension (.h) Renamed MarketWindow.cpp and marketWindow.h to lower case (keep it the same everywhere)
