+++ src/Widgets/marketwindow.h
